% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agt_coefficients.R
\docType{data}
\name{agt_coefficients}
\alias{agt_coefficients}
\title{Age Graded Tables}
\format{
## `agt_coefficients`
A data frame with 3'990 rows and 5 columns:
\describe{
  \item{gender}{gender, "F" or "M"}
  \item{age}{age}
  \item{coeff}{coefficient}
  \item{distance}{in kilometers}
  \item{record_sec}{record in secods}
  ...
}
}
\source{
<http://www.howardgrubb.co.uk/athletics/>
}
\usage{
agt_coefficients
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
