% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_to_best.R
\name{compare_to_best}
\alias{compare_to_best}
\title{Compare a Running Result}
\usage{
compare_to_best(mytime, distance, gender, age)
}
\arguments{
\item{mytime}{character. Time as HH:MM:SS}

\item{distance}{number (in kilometers)}

\item{gender}{character ("F" or "M")}

\item{age}{number (in years)}
}
\value{
A number (percentage of best in this distance/age/gender combination); 100 would be the world record.
}
\description{
Compare a Running Result
}
\examples{
compare_to_best("4:11:00", 42.190, "M", 51)  # Marathon
compare_to_best("0:54:33", 10.000, "F", 36)  # 10k
compare_to_best("1:38:15", 21.6648, "M", 40)  # Half marathon with some slope correction
}
