% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agt_grade.R
\name{convert_seconds_to_hms}
\alias{convert_seconds_to_hms}
\title{Time string to seconds}
\usage{
convert_seconds_to_hms(seconds)
}
\arguments{
\item{seconds}{number}
}
\description{
Time string to seconds
}
\keyword{internal}
