% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agt_grade.R
\name{get_agt_grade}
\alias{get_agt_grade}
\title{Compare a Running Result}
\usage{
get_agt_grade(mytime, distance, gender, age)
}
\arguments{
\item{mytime}{character. Time as HH:MM:SS}

\item{distance}{number (in kilometers)}

\item{gender}{character ("F" or "M")}

\item{age}{number (in years)}
}
\value{
A number (percentage of best in this distance/age/gender combination); 100 would be the world record.
}
\description{
Compare a Running Result
}
\examples{
get_agt_grade("4:11:00", 42.190, "M", 51) # Marathon
get_agt_grade("0:54:33", 10.000, "F", 36) # 10k
get_agt_grade("1:38:15", 21.6648, "M", 40) # Half marathon with some slope correction
}
