% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_records.R
\name{get_records}
\alias{get_records}
\title{Get Record Times for a Specific Combination of Age and Gender}
\usage{
get_records(gender, age)
}
\arguments{
\item{gender}{character ("F" or "M")}

\item{age}{number (in years)}
}
\value{
A data.frame (a subset of agt_coefficents with colums coeff, distance, record_sec)
}
\description{
Get Record Times for a Specific Combination of Age and Gender
}
\examples{
get_records("M", 51) # Marathon
}
