% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_to_best.R
\name{perform_interpolation}
\alias{perform_interpolation}
\title{Interpolate a specific column out of the agt_coefficients}
\usage{
perform_interpolation(distance, gender, age, column)
}
\arguments{
\item{distance}{number (in kilometers)}

\item{gender}{character ("F" or "M")}

\item{age}{number (in years)}

\item{column}{character ("record_sec" of "coeff")}
}
\description{
Interpolate a specific column out of the agt_coefficients
}
\keyword{internal}
