% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_to_best.R
\name{time_to_seconds}
\alias{time_to_seconds}
\title{Seconds from time string}
\usage{
time_to_seconds(time_str)
}
\arguments{
\item{time_str}{character. Time as HH:MM:SS}
}
\description{
Seconds from time string
}
\keyword{internal}
