% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agt_coefficients.R
\docType{data}
\name{agt_coefficients}
\alias{agt_coefficients}
\title{Age Graded Tables}
\format{
## `agt_coefficients`
A data frame with 3'990 rows and 5 columns:
\describe{
  \item{gender}{gender, "F" or "M"}
  \item{age}{age}
  \item{coeff}{coefficient}
  \item{distance}{in kilometers}
  \item{record_sec}{record in secods}
  ...
}
}
\source{
https://github.com/AlanLyttonJones/Age-Grade-Tables/tree/master/2020%20Files
}
\usage{
agt_coefficients
}
\description{
Compiled from the "2020 Files" at https://github.com/AlanLyttonJones/Age-Grade-Tables
Report ...
}
\keyword{datasets}
