% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agt_grade.R
\name{perform_interpolation}
\alias{perform_interpolation}
\title{Interpolate a specific column out of the agt_coefficients}
\usage{
perform_interpolation(distance, gender, age)
}
\arguments{
\item{distance}{number (in kilometers)}

\item{gender}{character ("F" or "M")}

\item{age}{number (in years)}
}
\value{
an array of 2: interpolated seconds, interpolated coefficient
}
\description{
Interpolate a specific column out of the agt_coefficients
}
\keyword{internal}
